requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
whatsapp-web>=0.0.1
qrcode-terminal>=0.8